+++ src/main/java/uniandes/isis2304/parranderos/persistencia/SQLEstante.java
@@ -47,7 +47,7 @@ class SQLEstante
 	 * @param pm - El manejador de persistencia
 	 * @return EL número de tuplas insertadas
 	 */
-	public long adicionarEstante (PersistenceManager pm, long idEstante, int espacio, long idBodega) 
+	public long adicionarEstante (PersistenceManager pm, long idEstante, long espacio, long idBodega) 
 	{
         Query q = pm.newQuery(SQL, "INSERT INTO " + pp.darTablaEstante () + "(id, espacio, bodega) values (?, ?, ?)");
         q.setParameters(idEstante, espacio, idBodega);
